Updated to Python scripts so that the installation should work on Windows.
Updated to the same requirements system as sire, specifying all requirements
in `requirements.txt` and `requirements_build.txt`. These are parsed
by `update_recipe.py` to create the conda recipe.

I have moved the sire version requirement into requirements.txt so that
it is easier to specify the sire version on which a pull request or
push depends. This way we don't need to be playing around with the
tags or anything. This also lets us control the range of supported
sire versions more easily, e.g. not having to use the major number
as the pin.

I have also updated the build process so that we don't need to
clone the sire or biosimspace repositories again - for the tests
we just copy the tests from the already-cloned repo.

Note that this won't build an run now because it depends
on sire >= 2023.0.0, < 2023.1.0