Use the Free Energy Analysis pipeline from alchemlyb (#24)
Use the Free Energy Analysis pipeline from alchemlyb. I'm sure that the tests could be abstracted more but I don't know how to do it.

Use the devel version of alchemlyb and will switch to the conda one when 1.0.0 is released.
Use the alchemlyb.workflow to perform analysis.
Write tests to check the new implementation.

Co-authored-by: Zhiyi Wu <zwu@exscientia.co.uk>