Add getVolume to Gromacs Process (#37)
This PR adds getVolume to Gromacs Process.

Since the volume is not present in the gromacs text output, this PR switched from reading from gromacs text output to use gmx energy to get all the energy terms.
Use python dictionary instead of the process._MultiDict() as I cannot figure out how to set a value instead of extend a list.
getStep is no longer reading from the log file but is computed based on time and timestep
When Unit is not given to getRecord. A float of the default unit will be returned.
When Unit is given to getRecord. The default unit will be returned. In the future I might want to change the output to match the input unit.
Add a TODO to convert constrained RMSD to unitless variable.

---------

Co-authored-by: Zhiyi Wu <zwu@exscientia.co.uk>