First pass at porting @annamherz's analysis code.
Lots of refactoring to remove code duplication and make
ongoing maintanence easier. GROMACS analysis appears to
work fine, although I am seeing very different values for
the TI estimator. SOMD analysis is currently failing due
to a spurious value in the input data. I'm not sure if
this is to do with my test data, or a bug from the port.