Remove GROMACS version check since meaningless in a staticmethod.
If this is important, then we can just raise a meaningful exception
if the alchemlyb analysis fails. We should be able to determine the
version using the header from the XVG file. In fact, the test files
were generated with GROMACS 2018, but still work fine with alchemlyb,
so the check doesn't appear to be necessary.