Implement multiple distance restraints for GROMACS
This is done by using restraint potentials (function type 10)
for the flat-bottomed restraints which are affected by lambda-
restraint, and a distance restraint for the permanent restraint
which is not affected by any lambdas. See
https://manual.gromacs.org/documentation/2019-rc1/reference-manual/topologies/topology-file-formats.html#id31
for details. Note that the distance restraint requires the force
constant to be written to the mdp file.